Begin ToDo implementation
Started putting together ToDo, the module for creating and managing
organized lists and for keeping track of projects. Also includes a silly
"give me a number" function, which thinks for a variable duration before
returning a number over 100.

The format for nested lists has been fully implemented (on backend).
Requires implementation of functions for user's adding and editing
lines.