Implemented program-wide exception handling!
About time. This unifies the interface for errors, making things more
consistent for the user and much easier to edit error messages. This
also fixes the CALCULATOR module from crashing the whole program on
invalid input.

Improvement to CALCULATOR help text is included, and miscellaneous code
comments added in a couple modules.